+++ tt/transformations/__init__.py
@@ -2,6 +2,7 @@
 
 from .bexpr import (  # noqa
     apply_de_morgans,
+    apply_identity_law,
     distribute_ands,
     distribute_ors,
     coalesce_negations,