+++ tt/trees/__init__.py
@@ -1,4 +1,10 @@
-"""Tools for working with Boolean expression trees."""
+"""Tools for working with Boolean expression trees.
+
+It should be noted that virtually all of the functionality within this module
+is presented with an easier-to-use interface in the :mod:`expressions \
+<tt.expressions>` module.
+
+"""
 
 from .tree_node import (  # noqa
     BinaryOperatorExpressionTreeNode,